backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
azure-ai-contentunderstanding>=1.0.0b1
azure-core>=1.30.0
